requirements.txt
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
Babel==2.10.3
bcrypt==4.1.3
blinker==1.6.2
certifi==2022.9.24
cffi==1.16.0
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.6
colorama==0.4.6
configobj==5.0.8
cryptography==38.0.4
databases==0.9.0
distro==1.8.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.4
httptools==0.6.1
httpx==0.27.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.2.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
jwt==1.3.1
keyring==24.2.0
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
mysql-connector-python==8.4.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
orjson==3.10.3
passlib==1.7.4
pexpect==4.8.0
Pillow==10.0.0
ptyprocess==0.7.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.0
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2023.3
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.3.1
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==1.26.16
uvicorn==0.29.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
xdg==5.0.0
zipp==1.0.0
